Wait for the parser to reach end of document when creating SOAP envelopes
Currently, when the SOAP body’s closing tag is reached by the StAXSOAPModelBuilder, it stops parsing the rest of the data.
In cases where the SOAP envelope’s closing tag is decoded into the buffer in the next iteration,
the decoder won’t complete the process because the parser has already stopped.
Therefore, we need to continue parsing the tokens until we reach the end-of-document token.